functions/pipeline-orchestrator/requirements.txt
azure-functions>=1.18.0,<2.0.0
azure-servicebus>=7.12.0,<8.0.0
azure-identity>=1.24.0,<2.0.0
